fix bug

+++ v1/1.0/pack/nart/src/main/java/com/nart/service/impl/CommentServiceImpl.java
@@ -2,7 +2,6 @@ package com.nart.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.nart.pojo.Comment;
-import com.nart.pojo.Status;
 import com.nart.pojo.User;
 import com.nart.service.CommentService;
 import com.nart.service.DataCounterService;
@@ -27,7 +26,7 @@ public class CommentServiceImpl implements CommentService {
     public List<Comment> showCommentList(String statusId) {
         LambdaQueryWrapper<Comment> lqw = new LambdaQueryWrapper<Comment>();
 
-        lqw.eq(Comment::getStatusId, Integer.valueOf(statusId));
+        lqw.eq(Comment::getStatusId, Long.valueOf(statusId));
         lqw.orderBy(true,false, Comment::getCreateDate);
 
         List<Comment> Comments = CommentDao.selectList(lqw);
